ml_service/app/db.py
from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession, async_sessionmaker
from sqlalchemy.orm import declarative_base
import os
from dotenv import load_dotenv

load_dotenv()

# Use the same DB as VIL Core
DATABASE_URL = os.getenv("DATABASE_URL", "postgresql+asyncpg://vil:vilpass@localhost:5432/vildb")

engine = create_async_engine(DATABASE_URL, echo=False)
async_session = async_sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)

Base = declarative_base()
