ml_service/requirements.txt
fastapi
uvicorn
httpx
pydantic-settings
sqlalchemy
asyncpg
pandas
scikit-learn
xgboost
joblib
matplotlib
seaborn
psycopg2-binary
python-dotenv
